+++ tectonic/dpx-dpxcrypt.c
@@ -98,7 +98,7 @@ static void transform (MD5_CONTEXT *ctx, const unsigned char *data)
   uint32_t *cwp = correct_words;
 
 #ifdef WORDS_BIGENDIAN
-  { int i; const unsigned char *p1; unsigned char *p2;
+  { unsigned int i; const unsigned char *p1; unsigned char *p2;
     for (i = 0, p1 = data, p2 = (unsigned char *)correct_words; i < 16; i++, p2 += 4 ) {
       p2[3] = *p1++; p2[2] = *p1++; p2[1] = *p1++; p2[0] = *p1++;
     }
@@ -1239,7 +1239,7 @@ AES_cbc_encrypt (const unsigned char *key,    size_t  key_len,
   const unsigned char *inptr;
   unsigned char *outptr, block[AES_BLOCKSIZE];
   size_t len;
-  int    i;
+  size_t i;
   int    padbytes;
 
   ctx = &aes;
@@ -1665,7 +1665,7 @@ static const uint32_t rcon[] =
  */
 int rijndaelSetupEncrypt(uint32_t *rk, const uint8_t *key, int keybits)
 {
-  int       i = 0;
+  unsigned int i = 0;
   uint32_t  temp;
 
   rk[0] = GETU32(key     );
