+++ tectonic/dpx-pdfobj.h
@@ -61,7 +61,7 @@ void     pdf_out_init      (const char *filename,
                                    bool enable_encrypt, bool enable_object_stream);
 void     pdf_out_flush     (void);
 void     pdf_set_version   (unsigned version);
-unsigned pdf_get_version   (void);
+unsigned int pdf_get_version   (void);
 
 void     pdf_release_obj (pdf_obj *object);
 int      pdf_obj_typeof  (pdf_obj *object);
@@ -94,10 +94,10 @@ pdf_obj *pdf_new_number     (double value);
 void     pdf_set_number     (pdf_obj *object, double value);
 double   pdf_number_value   (pdf_obj *number);
 
-pdf_obj  *pdf_new_string    (const void *str, unsigned length);
-void      pdf_set_string    (pdf_obj *object, unsigned char *str, unsigned length);
+pdf_obj  *pdf_new_string    (const void *str, size_t length);
+void      pdf_set_string    (pdf_obj *object, unsigned char *str, size_t length);
 void     *pdf_string_value  (pdf_obj *object);
-unsigned  pdf_string_length (pdf_obj *object);
+unsigned int  pdf_string_length (pdf_obj *object);
 
 /* Name does not include the / */
 pdf_obj *pdf_new_name   (const char *name);
@@ -111,7 +111,7 @@ pdf_obj *pdf_new_array     (void);
  */
 void     pdf_add_array     (pdf_obj *array, pdf_obj *object);
 pdf_obj *pdf_get_array     (pdf_obj *array, int idx);
-unsigned pdf_array_length  (pdf_obj *array);
+unsigned int pdf_array_length  (pdf_obj *array);
 
 pdf_obj *pdf_new_dict    (void);
 void     pdf_remove_dict (pdf_obj *dict,  const char *key);
@@ -174,13 +174,13 @@ int       check_for_pdf     (rust_input_handle_t handle);
 pdf_file *pdf_open          (const char *ident, rust_input_handle_t handle);
 void      pdf_close         (pdf_file *pf);
 pdf_obj  *pdf_file_get_trailer (pdf_file *pf);
-int       pdf_file_get_version (pdf_file *pf);
+unsigned int  pdf_file_get_version (pdf_file *pf);
 pdf_obj  *pdf_file_get_catalog (pdf_file *pf);
 
 pdf_obj *pdf_deref_obj     (pdf_obj *object);
 pdf_obj *pdf_import_object (pdf_obj *object);
 
-int      pdfobj_escape_str (char *buffer, int size, const unsigned char *s, int len);
+size_t pdfobj_escape_str (char *buffer, size_t size, const unsigned char *s, size_t len);
 
 pdf_obj *pdf_new_indirect  (pdf_file *pf, unsigned label, unsigned short generation);
